some refactoring, renaming for clarity
